feat(#469): extract shared project table row, card, and actions into separate components
